refactoring code in Header
