+++ src/common/header/Header.jsx
@@ -1,49 +1,23 @@
 import React, { useRef, useEffect } from "react";
 import { Link, useLocation } from "react-router-dom";
 
-import ScrollToTop from "../scroll-to-top/ScrollToTop";
+import ScrollToTop from "../../components/scroll-to-top/ScrollToTop";
+
+import shrinkHeader from "./helpers/shrinkHeader";
+import headerNav from "./helpers/headerNav";
+
 import "./header.scss";
-
 import logo from "../../assets/tmovie.png";
-
-const headerNav = [
-  {
-    display: "Home",
-    path: "/",
-    id: 1,
-  },
-  {
-    display: "Movies",
-    path: "/movie",
-    id: 2,
-  },
-  {
-    display: "TV Series",
-    path: "/tv",
-    id: 3,
-  },
-];
 
 const Header = () => {
   const { pathname } = useLocation();
   const headerRef = useRef(null);
-
   const active = headerNav.findIndex((e) => e.path === pathname);
 
   useEffect(() => {
-    const shrinkHeader = () => {
-      if (
-        document.body.scrollTop > 100 ||
-        document.documentElement.scrollTop > 100
-      ) {
-        headerRef.current.classList.add("shrink");
-      } else {
-        headerRef.current.classList.remove("shrink");
-      }
-    };
-    window.addEventListener("scroll", shrinkHeader);
+    window.addEventListener("scroll", () => shrinkHeader(headerRef));
     return () => {
-      window.removeEventListener("scroll", shrinkHeader);
+      window.removeEventListener("scroll", () => shrinkHeader(headerRef));
     };
   }, []);
