+++ src/common/header/helpers/headerNav.js
@@ -1,0 +1,19 @@
+const headerNav = [
+  {
+    display: "Home",
+    path: "/",
+    id: 1,
+  },
+  {
+    display: "Movies",
+    path: "/movie",
+    id: 2,
+  },
+  {
+    display: "TV Series",
+    path: "/tv",
+    id: 3,
+  },
+];
+
+export default headerNav;
