+++ src/common/header/helpers/shrinkHeader.js
@@ -1,0 +1,12 @@
+const shrinkHeader = (headerRef) => {
+  if (
+    document.body.scrollTop > 100 ||
+    document.documentElement.scrollTop > 100
+  ) {
+    headerRef.current.classList.add("shrink");
+  } else {
+    headerRef.current.classList.remove("shrink");
+  }
+};
+
+export default shrinkHeader;
